refactor(cloister): extract specialists lifecycle seam (pan-2232)
